requirements.txt
alembic==1.12.1
amqp==5.2.0
attrs==23.1.0
billiard==4.2.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
deprecation==2.1.0
docker==6.1.3
exceptiongroup==1.1.3
Flask==2.2.2
Flask-CeleryExt==0.4.3
Flask-Cors==3.0.10
Flask-Migrate==4.0.0
flask-oidc==1.4.0
Flask-SQLAlchemy==3.0.2
flower==1.2.0
gunicorn==20.1.0
httplib2==0.22.0
humanize==4.8.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.3
Mako==1.2.4
MarkupSafe==2.1.3
oauth2client==4.1.3
packaging==23.2
pluggy==1.3.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
pytest==7.2.2
pytest-mock==3.10.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.45
testcontainers==3.7.1
tomli==2.0.1
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.9
websocket-client==1.6.4
Werkzeug==2.3.7
wrapt==1.15.0
zipp==3.17.0



